Replace `PlotAudit` trait with `PlotAudit` struct in farmer, no logic changes, just moving code around
